Add simple testing infrastructure
This is *not* our long term testing infrastructure. That will be self-hosted in
JS so that we don't need to start a fresh Starling process for every single
test.

This is just an initial sanity testing infrastructure so that we can start
building all the stuff needed so that we can start self-hosting the testing
infrastructure.