Gluing Rust's `Future` and JavaScript's `Promise` together.
JavaScript's `Promise` and Rust's `Future` are two abstractions with the same
goal: asynchronous programming with eventual values. Both `Promise`s and
`Future`s represent a value that may or may not have been computed yet. However,
the way that you interact with each of them is very different.

JavaScript's `Promise` follows a completion-based model. You register callbacks
on a promise, and the runtime invokes each of the registered callbacks (there
may be many!) when the promise is resolved with a value. You build larger
asynchronous computations by chaining promises; registering a callback with a
promise returns a new promise of that callback's result. Dependencies between
promises are managed by the runtime.

Rust's `Future` follows a readiness-based model. You define a `poll` method that
either returns the future's value if it is ready, or a sentinel that says the
future's value is not ready yet. You build larger asynchronous computations by
defining (or using existing) combinators that wrap some inner future (or
futures). These combinators delegate polling to the inner future, and once the
inner future's value is ready, perform their computation on the value, and
return the transformed result. Dependencies between futures are managed by the
futures themselves, while scheduling polling is the runtime's responsibility.

To translate between futures and promises we take two different approaches for
each direction, by necessity.

To treat a Rust `Future` as a JavaScript `Promise`, we define a `Future`
combinator called `Future2Promise`. It takes a fresh, un-resolved `Promise`
object, and an inner future upon construction. It's `poll` method delegates to
the inner future's `poll`, and once the inner future's value is ready, it
resolves the promise with the ready value, and the JavaScript runtime ensures
that the promise's registered callbacks are invoked appropriately.

To treat a JavaScript `Promise` as a Rust `Future`, we register a callback to
the promise that sends the resolution value over a one-shot channel. One-shot
channels are split into their two halves: sender and receiver. The sender half
moves into the callback, but the receiver half is a future, and it represents
the future resolution value of the original promise.

The final concern to address is that the runtime scheduling the polling of Rust
`Future`s (`tokio`) still knows which futures to poll despite it only seeing
half the picture now. I say "half the picture" because dependencies that would
otherwise live fully within the futures ecosystem are now hidden in promises
inside the JavaScript runtime.

First, we must understand how `tokio` schedules polling. It is not busy spinning
and calling `poll` continuously in a loop. `tokio` maintains a set of "root"
futures. These are the futures passed to `Core::run` and `Handle::spawn`
directly. When `tokio` polls a "root" future, that `poll` call will transitively
reach down and call `poll` on "leaf" futures that wrap file descriptors and
sockets and such things. It is these "leaf" futures' responsibilty to use OS
APIs to trigger wake ups when new data is available on a socket or what have
you, and then it is `tokio`'s responsibilty to map that wake up back to which
"root" future it should poll again. If the "leaf" futures do not properly
register to be woken up again, `tokio` will never poll that "root" future again,
effectively dead locking it.

So we must ensure that our `Promise`-backed futures will always be polled again
by making sure that they have proper "leaf" futures. Luckily, the receiver half
of a one-shot channel is such a "leaf" future that properly registers future
wake ups. If instead, for example, we tried directly checking the promise's
state in `poll` with JSAPI methods, we *wouldn't* register any wake ups, `tokio`
would never `poll` the future again, and the future would dead lock.