Add support for simple timeouts
This commit uses the `tokio_timer` crate to get futures that become ready when a
given amount of time has passed. These are exposed to JavaScript via the
`timeout` function.

This also adds the behavior where we execute the `main` JS function and then
shut down the task after its promise resolves.