Add the ability to spawn new tasks
After this commit, JS tasks can spawn children JS tasks. Spawning a child task
returns a promise that is resolved when the child exits OK and rejected when the
child exits due to some kind of failure.

A parent and child task's fates are forever linked:

* If a parent task exits, then the child is task is also killed.

* If a child task exits due to an unhandled error or rejected promise, then the
  promise representing that child task is rejected, and the parent has one event
  loop tick to handle that promise rejection before the failure cascades up to
  the grandparent.

The API for spawning new tasks is currently `spawn("path/to/file/js")` where the
given path is relative to the JS file executing this current task. A promise
representing the child's lifetime is returned.

Eventually, we plan to also set up an initial readable and writable stream pair
for communication between the child and parent, as well as a mechanism for the
parent to force-kill the child. Additionally, we'd like to allow spawning
JS *modules* rather than *files*, but that is not implemented yet either.